fixed Ali's feedback
